Moved 'config related' classes/files to a sub-directory
