+++ srcs/config/AConfig.cpp
@@ -1,4 +1,4 @@
-#include <AConfig.hpp>
+#include <config/AConfig.hpp>
 #include <Logger.hpp>
 
 AConfig::AConfig(const map_type& map) : _vars(map)
